3.0.0.22
Fix for #3432, #3434

+++ install/install.queries.php
@@ -1226,7 +1226,7 @@ $SETTINGS = array (';
                         $mysqli_result = mysqli_query(
                             $dbTmp,
                             "CREATE TABLE IF NOT EXISTS `" . $var['tbl_prefix'] . "cache_tree` (
-                            `increment_id` tinyint(32) NOT NULL AUTO_INCREMENT,
+                            `increment_id` smallint(32) NOT NULL AUTO_INCREMENT,
                             `data` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_bin NOT NULL CHECK (json_valid(`data`)),
                             `visible_folders` longtext NOT NULL,
                             `timestamp` varchar(50) NOT NULL,
